FST 83 - move some files
